refactor(workstation): extract 4 simple surfaces from inkRuntime (#890 phase 5a.1)
First batch of surface extractions following the foundation in #896
(phase 5a.0). Pulls 4 of the 12 promoted-view surfaces out of
inkRuntime.ts into per-surface modules under src/workstation/surfaces/.

Surfaces moved (each ~80-180 LOC):
  - reflog       (#781) → src/workstation/surfaces/reflog/index.ts
  - bisect       (#784) → src/workstation/surfaces/bisect/index.ts
  - stash        → src/workstation/surfaces/stash/index.ts
  - worktrees    → src/workstation/surfaces/worktrees/index.ts

Each is a near-verbatim cut. The only call-site change is the inline
`truncate` alias becoming `truncateCells` (already imported from
chrome/text); no rendering output changes.

New shared module:
  - src/workstation/runtime/promotedFilter.ts (52 LOC) — holds the
    two helpers that every promoted surface uses: matchesPromotedFilter
    (case-insensitive substring across haystacks) and
    renderPromotedFilterAffordance (the inline filter input). Lives in
    runtime/ rather than chrome/ because both are tightly coupled to the
    LogInkState filter-mode shape.

inkRuntime.ts: 5,879 → 5,485 LOC (-394 this PR; -554 cumulative since
phase 5a.0). Removes 4 surface render functions, 2 promoted-filter
helpers, and 3 imports that became unused after extraction
(formatLogInkReflogEmpty, formatLogInkStashEmpty, splitReflogSubject).

Why these 4 first: each is self-contained — no closure dependencies on
LogInkApp's local hook state, no inter-surface helpers, just a render
function that takes (state, context, theme, layout) and returns Ink
nodes. Validates the extraction pattern before tackling the more
intricate surfaces (history, status, diff, detail) in 5a.2-5a.4.

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same numbers as post-5a.0 baseline). No behavior change —
every surface renders byte-identically before/after.
